+++ src/main/java/me/marin/lockout/mixin/server/WolfEntityMixin.java
@@ -2,8 +2,12 @@ package me.marin.lockout.mixin.server;
 
 import me.marin.lockout.Lockout;
 import me.marin.lockout.lockout.Goal;
+import me.marin.lockout.lockout.goals.misc.AngerWolfGoal;
 import me.marin.lockout.lockout.goals.wear_armor.PutWolfArmorOnWolfGoal;
 import me.marin.lockout.server.LockoutServer;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.animal.wolf.Wolf;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.InteractionResult;
@@ -15,6 +19,27 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(Wolf.class)
 public class WolfEntityMixin {
+
+    @Inject(method = "hurtServer", at = @At("RETURN"))
+    public void onWolfHurt(ServerLevel level, DamageSource damageSource, float amount, CallbackInfoReturnable<Boolean> cir) {
+        if (!cir.getReturnValue()) return;
+        Wolf wolf = (Wolf)(Object) this;
+        if (!wolf.isAngry()) return;
+        if (!(damageSource.getEntity() instanceof ServerPlayer player)) return;
+
+        Lockout lockout = LockoutServer.lockout;
+        if (!Lockout.isLockoutRunning(lockout)) return;
+
+        for (Goal goal : lockout.getBoard().getGoals()) {
+            if (goal == null) continue;
+            if (goal.isCompleted()) continue;
+
+            if (goal instanceof AngerWolfGoal) {
+                lockout.completeGoal(goal, player);
+                return;
+            }
+        }
+    }
 
     @Inject(method = "mobInteract", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/animal/wolf/Wolf;setItemSlotAndDropWhenKilled(Lnet/minecraft/world/entity/EquipmentSlot;Lnet/minecraft/world/item/ItemStack;)V", ordinal = 0))
     public void onEquipArmor(Player player, InteractionHand hand, CallbackInfoReturnable<InteractionResult> cir) {